requirements.txt
beautifulsoup4
selenium
webdriver_manager
python-dotenv